Refactor stream framing code
